feat: executor level、commissionRatio

+++ contracts/interfaces/IFeeCollector.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.0;
 interface IFeeCollector {
 
     event UpdateLevelDiscountRatio(
-        uint256 level,
+        uint8 level,
         uint256 oldRatio,
         uint256 newRatio
     );
@@ -14,11 +14,11 @@ interface IFeeCollector {
         uint256 newRatio
     );
 
-    function levelDiscountRatios(uint256 level) external view returns (uint256 discountRatio);
+    function levelDiscountRatios(uint8 level) external view returns (uint256 discountRatio);
 
     function maxCommissionRatio() external view returns (uint256 maxCommissionRatio);
 
-    function updateLevelDiscountRatio(uint256 level, uint256 newRatio) external;
+    function updateLevelDiscountRatio(uint8 level, uint256 newRatio) external;
 
     function updateMaxCommissionRatio(uint256 newRatio) external;
 }
